final_project/requirements.txt
allure-pytest==2.13.2
pytest==7.4.0
requests==2.31.0
selenium==4.11.2
cerberus==1.3.4
